extract keyboard hook and pinball overlay
